finish api search weather from city

+++ app/src/main/java/com/example/weatherapp19052023/data/api/dto/MainFromCityDTO.kt
@@ -3,17 +3,17 @@ package com.example.weatherapp19052023.data.api.dto
 import com.google.gson.annotations.SerializedName
 
 data class MainFromCityDTO(
-    val temp: Long,
+    val temp: Double,
     @SerializedName("feels_like")
     val feelsLike: Double,
     @SerializedName("temp_min")
-    val tempMin: Long,
+    val tempMin: Double,
     @SerializedName("temp_max")
-    val tempMax: Long,
-    val pressure: Long,
-    val humidity: Long,
+    val tempMax: Double,
+    val pressure: Int,
+    val humidity: Int,
     @SerializedName("sea_level")
-    val seaLevel: Long,
+    val seaLevel: Int,
     @SerializedName("grnd_level")
-    val grndLevel: Long,
+    val grndLevel: Int,
 )
